Correccion nombres de paquetes en donde estan ubicadas las entidades
