+++ src/main/java/co/edu/udistrital/mdp/eventos/entities/bookingentity/NotificationEntity.java
@@ -1,7 +1,7 @@
 package co.edu.udistrital.mdp.eventos.entities.bookingentity;
 
 import co.edu.udistrital.mdp.eventos.entities.BaseEntity;
-import co.edu.udistrital.mdp.eventos.entities.userEntitys.Assistantentitys;
+import co.edu.udistrital.mdp.eventos.entities.userentity.Assistantentitys;
 import jakarta.persistence.Entity;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.OneToOne;
