--- src/main/java/co/edu/udistrital/mdp/eventos/entities/eventEntitys/LocationEntity.java
@@ -1,22 +1,0 @@
-package co.edu.udistrital.mdp.eventos.entities.eventEntitys;
-
-
-import co.edu.udistrital.mdp.eventos.entities.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.OneToOne;
-import lombok.Data;
-import uk.co.jemos.podam.common.PodamExclude;
-
-@Data
-@Entity
-
-public class LocationEntity extends BaseEntity{
-    @PodamExclude
-    @OneToOne
-    private EventEntity event;
-
-    String name;
-    String location;
-    String type;
-    Integer capacity;
-}
