+++ src/main/java/co/edu/udistrital/mdp/eventos/entities/paymententity/CardEntity.java
@@ -1,4 +1,4 @@
-package co.edu.udistrital.mdp.eventos.entities.paymentEntitys;
+package co.edu.udistrital.mdp.eventos.entities.paymententity;
 
 import java.util.Date;
