+++ src/main/java/co/edu/udistrital/mdp/eventos/entities/userentity/OrganizerEntity.java
@@ -1,10 +1,10 @@
-package co.edu.udistrital.mdp.eventos.entities.userEntitys;
+package co.edu.udistrital.mdp.eventos.entities.userentity;
 
 
 import java.util.ArrayList;
 import java.util.List;
 
-import co.edu.udistrital.mdp.eventos.entities.eventEntitys.EventEntity;
+import co.edu.udistrital.mdp.eventos.entities.evententity.EventEntity;
 import jakarta.persistence.Entity;
 import jakarta.persistence.OneToMany;
 import lombok.Data;
