+++ src/main/java/co/edu/udistrital/mdp/eventos/repositories/AssistantRepository.java
@@ -2,7 +2,7 @@ package co.edu.udistrital.mdp.eventos.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import co.edu.udistrital.mdp.eventos.entities.userEntitys.Assistantentitys;
+import co.edu.udistrital.mdp.eventos.entities.userentity.Assistantentitys;
 
 public interface AssistantRepository extends JpaRepository<Assistantentitys, Long>{
