+++ src/main/java/co/edu/udistrital/mdp/eventos/repositories/CreditCardRepository.java
@@ -2,7 +2,7 @@ package co.edu.udistrital.mdp.eventos.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import co.edu.udistrital.mdp.eventos.entities.paymentEntitys.CreditCardEntity;
+import co.edu.udistrital.mdp.eventos.entities.paymententity.CreditCardEntity;
 
 public interface CreditCardRepository extends JpaRepository<CreditCardEntity,Long>{
