+++ src/main/java/co/edu/udistrital/mdp/eventos/repositories/EventRespository.java
@@ -1,7 +1,7 @@
 package co.edu.udistrital.mdp.eventos.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import co.edu.udistrital.mdp.eventos.entities.eventEntitys.EventEntity;
+import co.edu.udistrital.mdp.eventos.entities.evententity.EventEntity;
 
 public interface EventRespository extends JpaRepository<EventEntity, Long> {
