+++ src/main/java/co/edu/udistrital/mdp/eventos/repositories/LocationRespository.java
@@ -1,7 +1,7 @@
 package co.edu.udistrital.mdp.eventos.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import co.edu.udistrital.mdp.eventos.entities.eventEntitys.LocationEntity;
+import co.edu.udistrital.mdp.eventos.entities.evententity.LocationEntity;
 
 public interface LocationRespository extends JpaRepository<LocationEntity, Long> {
