+++ src/main/java/co/edu/udistrital/mdp/eventos/repositories/MobileWalletRepository.java
@@ -2,7 +2,7 @@ package co.edu.udistrital.mdp.eventos.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import co.edu.udistrital.mdp.eventos.entities.paymentEntitys.MobileWalleteEntity;
+import co.edu.udistrital.mdp.eventos.entities.paymententity.MobileWalleteEntity;
 
 public interface MobileWalletRepository extends JpaRepository<MobileWalleteEntity, Long>{
