+++ src/main/java/co/edu/udistrital/mdp/eventos/repositories/OrganizerRepository.java
@@ -2,7 +2,7 @@ package co.edu.udistrital.mdp.eventos.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import co.edu.udistrital.mdp.eventos.entities.userEntitys.OrganizerEntity;
+import co.edu.udistrital.mdp.eventos.entities.userentity.OrganizerEntity;
 
 public interface OrganizerRepository extends JpaRepository<OrganizerEntity, Long>{
