+++ src/main/java/co/edu/udistrital/mdp/eventos/repositories/PreferenceRepository.java
@@ -2,7 +2,7 @@ package co.edu.udistrital.mdp.eventos.repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import co.edu.udistrital.mdp.eventos.entities.userEntitys.PreferenceEntity;
+import co.edu.udistrital.mdp.eventos.entities.userentity.PreferenceEntity;
 
 public interface PreferenceRepository extends JpaRepository<PreferenceEntity, Long> {
