+++ src/main/java/co/edu/udistrital/mdp/eventos/repositories/ResourceRespository.java
@@ -1,7 +1,7 @@
 package co.edu.udistrital.mdp.eventos.repositories;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import co.edu.udistrital.mdp.eventos.entities.eventEntitys.ResourceEntity;
+import co.edu.udistrital.mdp.eventos.entities.evententity.ResourceEntity;
 
 public interface ResourceRespository extends JpaRepository<ResourceEntity, Long> {
